requirements.txt
# Core LangChain & LLM
langchain==0.2.14
langchain-community==0.2.12
langchain-groq==0.1.9

# Data parsing / processing
pandas
openpyxl==3.1.5
PyPDF2==3.0.1
python-docx==1.1.2

# App frontend
streamlit==1.35.0

# Web scraping / text cleaning
unstructured>=0.14.6
selenium==4.21.0

# Vector DB (if used by portfolio)
chromadb==0.5.0

# Environment + API utilities
python-dotenv==1.0.0

# Google Mail API
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.151.0

# Optional but recommended
tqdm==4.66.5
requests==2.32.3
lxml==5.3.0
typing_extensions==4.12.2
